requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
django=3.1.2=py_0
django-appconf=1.0.4=pyh9f0ad1d_0
django-extensions=3.1.5=pyhd8ed1ab_0
djangorestframework=3.13.1=pyhd8ed1ab_0
idna=3.3=pyhd8ed1ab_0
intel-openmp=2021.4.0=hecd8cb5_3538
mysql-connector-c=6.1.11=h0f02589_1007
mysqlclient=2.0.3=py37h23ab428_1
ncurses=6.3=hca72f7f_2
numexpr=2.8.1=py37h2e5f0a9_0
numpy=1.21.5=py37h2e5f0a9_2
numpy-base=1.21.5=py37h3b1a694_2
oauthlib=3.2.0=pyhd8ed1ab_0
openssl=1.1.1o=hfe4f2af_0
packaging=21.3=pyhd3eb1b0_0
pip=21.2.2=py37hecd8cb5_0
psycopg2=2.8.5=py37hddc9c9b_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.3.0=pyhd8ed1ab_1
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py37hf985489_4
python=3.7.11=h88f2d9e_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python3-openid=3.2.0=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytz=2020.1=py_0
readline=8.1.2=hca72f7f_1
requests=2.27.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
setuptools=58.0.4=py37hecd8cb5_0
shortuuid=1.0.8=py37hf985489_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.37.2=h707629a_0
sqlparse=0.4.1=py_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h7bc2e8c_0
urllib3=1.26.8=pyhd8ed1ab_1
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h4dc903c_4
